cleanup

+++ FlipperZero/game/loading.hpp
@@ -11,14 +11,14 @@ public:
     //
     void setText(const char *text) { currentText = text; }
     //
-    uint16_t getTimeElapsed() { return timeElapsed; }
+    uint32_t getTimeElapsed() { return timeElapsed; }
 
 private:
     Draw *draw = nullptr;
     void drawSpinner();
-    uint16_t spinnerPosition;
-    uint16_t timeElapsed;
-    uint16_t timeStart;
+    uint32_t spinnerPosition;
+    uint32_t timeElapsed;
+    uint32_t timeStart;
     bool animating = false;
     const char *currentText = "Loading...";
 };
